Use double precision Jitter in JitterDemo

+++ src/JitterDemo/Demos/Car/RayCastCar.cs
@@ -44,36 +44,36 @@ public class RayCastCar
 
     public RigidBody Body { get; }
 
-    private float destSteering;
-    private float destAccelerate;
-    private float steering;
-    private float accelerate;
+    private double destSteering;
+    private double destAccelerate;
+    private double steering;
+    private double accelerate;
 
     /// <summary>
     /// The maximum steering angle in degrees
     /// for both front wheels
     /// </summary>
-    public float SteerAngle { get; set; }
+    public double SteerAngle { get; set; }
 
     /// <summary>
     /// The maximum torque which is applied to the
     /// car when accelerating.
     /// </summary>
-    public float DriveTorque { get; set; }
+    public double DriveTorque { get; set; }
 
     /// <summary>
     /// Lower/Higher the acceleration of the car.
     /// </summary>
-    public float AccelerationRate { get; set; }
+    public double AccelerationRate { get; set; }
 
     /// <summary>
     /// Lower/Higher the steering rate of the car.
     /// </summary>
-    public float SteerRate { get; set; }
+    public double SteerRate { get; set; }
 
     // don't damp perfect, allow some bounciness.
-    private const float dampingFrac = 0.8f;
-    private const float springFrac = 0.45f;
+    private const double dampingFrac = 0.8d;
+    private const double springFrac = 0.45d;
 
     /// <summary>
     /// Initializes a new instance of the DefaultCar class.
@@ -86,39 +86,39 @@ public class RayCastCar
 
         // set some default values
         AccelerationRate = 10f;
-        SteerAngle = (float)JAngle.FromDegree(40.0f);
-        DriveTorque = 340.0f;
-        SteerRate = 5.0f;
+        SteerAngle = (double)JAngle.FromDegree(40.0d);
+        DriveTorque = 340.0d;
+        SteerRate = 5.0d;
 
         Body = world.CreateRigidBody();
 
-        TransformedShape tfs1 = new(new BoxShape(3.1f, 1.4f, 8f), new JVector(0, 0.6f, 0));
-        TransformedShape tfs2 = new(new BoxShape(2.4f, 0.8f, 5f), new JVector(0.0f, 1.7f, 1.1f));
+        TransformedShape tfs1 = new(new BoxShape(3.1d, 1.4d, 8f), new JVector(0, 0.6d, 0));
+        TransformedShape tfs2 = new(new BoxShape(2.4d, 0.8d, 5f), new JVector(0.0d, 1.7d, 1.1d));
 
         Body.AddShape(tfs1);
         Body.AddShape(tfs2);
 
-        float mass = 100.0f;
-        JVector sides = new JVector(3.1f, 1.0f, 8.0f);
+        double mass = 100.0d;
+        JVector sides = new JVector(3.1d, 1.0d, 8.0d);
 
-        float Ixx = (1.0f / 12.0f) * mass * (sides.Y * sides.Y + sides.Z * sides.Z);
-        float Iyy = (1.0f / 12.0f) * mass * (sides.X * sides.X + sides.Z * sides.Z);
-        float Izz = (1.0f / 12.0f) * mass * (sides.X * sides.X + sides.Y * sides.Y);
+        double Ixx = (1.0d / 12.0d) * mass * (sides.Y * sides.Y + sides.Z * sides.Z);
+        double Iyy = (1.0d / 12.0d) * mass * (sides.X * sides.X + sides.Z * sides.Z);
+        double Izz = (1.0d / 12.0d) * mass * (sides.X * sides.X + sides.Y * sides.Y);
 
         JMatrix inertia = new JMatrix(Ixx, 0, 0, 0, Iyy, 0, 0, 0, Izz);
         JVector r = new JVector(0, 0, 0);
         inertia += mass * r.LengthSquared() * JMatrix.Identity - mass * JVector.Outer(r, r);
 
-        Body.Position = new JVector(0, 0.5f, -4);
+        Body.Position = new JVector(0, 0.5d, -4);
         Body.SetMassInertia(inertia, mass);
 
-        Body.Damping = (0.0001f, 0.0001f);
+        Body.Damping = (0.0001d, 0.0001d);
 
         // create default wheels
-        Wheels[0] = new Wheel(world, Body, new JVector(-1.3f, 0.1f, -2.5f), 0.60f);
-        Wheels[1] = new Wheel(world, Body, new JVector(+1.3f, 0.1f, -2.5f), 0.60f);
-        Wheels[2] = new Wheel(world, Body, new JVector(-1.3f, 0.1f, +2.4f), 0.60f);
-        Wheels[3] = new Wheel(world, Body, new JVector(+1.3f, 0.1f, +2.4f), 0.60f);
+        Wheels[0] = new Wheel(world, Body, new JVector(-1.3d, 0.1d, -2.5d), 0.60d);
+        Wheels[1] = new Wheel(world, Body, new JVector(+1.3d, 0.1d, -2.5d), 0.60d);
+        Wheels[2] = new Wheel(world, Body, new JVector(-1.3d, 0.1d, +2.4d), 0.60d);
+        Wheels[3] = new Wheel(world, Body, new JVector(+1.3d, 0.1d, +2.4d), 0.60d);
 
         AdjustWheelValues();
     }
@@ -130,14 +130,14 @@ public class RayCastCar
     /// </summary>
     public void AdjustWheelValues()
     {
-        float mass = Body.Mass / 4.0f;
-        float wheelMass = Body.Mass * 0.03f;
+        double mass = Body.Mass / 4.0d;
+        double wheelMass = Body.Mass * 0.03d;
 
         foreach (Wheel w in Wheels)
         {
-            w.Inertia = 0.5f * (w.Radius * w.Radius) * wheelMass;
+            w.Inertia = 0.5d * (w.Radius * w.Radius) * wheelMass;
             w.Spring = mass * world.Gravity.Length() / (w.WheelTravel * springFrac);
-            w.Damping = 2.0f * (float)Math.Sqrt(w.Spring * Body.Mass) * 0.25f * dampingFrac;
+            w.Damping = 2.0d * (double)Math.Sqrt(w.Spring * Body.Mass) * 0.25d * dampingFrac;
         }
     }
 
@@ -159,36 +159,36 @@ public class RayCastCar
     /// the maximum steer angle by setting <see cref="SteerAngle" />. The speed of steering
     /// change is adjusted by <see cref="SteerRate" />.
     /// </param>
-    public void SetInput(float accelerate, float steer)
+    public void SetInput(double accelerate, double steer)
     {
         destAccelerate = accelerate;
         destSteering = steer;
     }
 
-    public void Step(float timestep)
+    public void Step(double timestep)
     {
         foreach (Wheel w in Wheels) w.PreStep(timestep);
 
-        float deltaAccelerate = timestep * AccelerationRate;
+        double deltaAccelerate = timestep * AccelerationRate;
 
-        float deltaSteering = timestep * SteerRate;
+        double deltaSteering = timestep * SteerRate;
 
-        float dAccelerate = destAccelerate - accelerate;
+        double dAccelerate = destAccelerate - accelerate;
         dAccelerate = Math.Clamp(dAccelerate, -deltaAccelerate, deltaAccelerate);
 
-        float dSteering = destSteering - steering;
+        double dSteering = destSteering - steering;
         dSteering = Math.Clamp(dSteering, -deltaSteering, deltaSteering);
 
         accelerate += dAccelerate;
         steering += dSteering;
 
-        float maxTorque = DriveTorque * 0.5f;
+        double maxTorque = DriveTorque * 0.5d;
 
         foreach (Wheel w in Wheels)
         {
             w.AddTorque(maxTorque * accelerate);
 
-            if (destAccelerate == 0.0f && w.AngularVelocity < 0.8f)
+            if (destAccelerate == 0.0d && w.AngularVelocity < 0.8d)
             {
                 // if the car is slow enough and destAccelerate is zero
                 // apply torque in the opposite direction of the angular velocity
@@ -197,7 +197,7 @@ public class RayCastCar
             }
         }
 
-        float alpha = SteerAngle * steering;
+        double alpha = SteerAngle * steering;
 
         Wheels[0].SteerAngle = alpha;
         Wheels[1].SteerAngle = alpha;
